task/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.8.6
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.4
sqlparse==0.4.3
python-jose==3.3.0